requirements.txt
fastapi
uvicorn
yt-dlp
openai
ffmpeg-python
pydantic
